services/data_augmentation/requirements.txt
alembic==1.16.4
annotated-types==0.7.0
antropy==0.1.9
anyio==4.11.0
arch==7.2.0
astroid==3.3.11
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.3.0
cloudpickle==3.1.1
contourpy==1.3.2
ctgan==0.11.0
cycler==0.12.1
databricks-sdk==0.62.0
dill==0.4.0
docker==7.1.0
einops==0.8.1
exceptiongroup==1.3.0
Faker==37.8.0
fastapi==0.118.0
filelock==3.18.0
Flask==3.1.1
fonttools==4.59.0
fsspec==2025.7.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.40.3
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.2.3
gunicorn==23.0.0
h11==0.16.0
h5py==3.14.0
hf-xet==1.1.7
huggingface-hub==0.34.3
idna==3.10
imbalanced-learn==0.14.0
imblearn==0.0
importlib_metadata==8.7.0
isort==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.8
llvmlite==0.44.0
Mako==1.3.10
mamba-ssm==2.2.5
Markdown==3.8.2
MarkupSafe==3.0.2
matplotlib==3.10.5
mccabe==0.7.0
mlflow==2.20.0
mlflow-skinny==2.20.0
mpmath==1.3.0
networkx==3.4.2
ninja==1.11.1.4
numba==0.61.2
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
opentelemetry-api==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
packaging==24.2
pandas==2.3.1
patsy==1.0.1
pillow==11.3.0
platformdirs==4.4.0
protobuf==5.29.5
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
pylint==3.3.8
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
PyWavelets==1.8.0
PyYAML==6.0.2
rdt==1.18.1
regex==2025.7.34
requests==2.32.4
rsa==4.9.1
safetensors==0.6.1
scikit-learn==1.7.1
scipy==1.15.3
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.42
sqlparse==0.5.3
starlette==0.48.0
statsmodels==0.14.5
stumpy==1.13.0
supersmoother==0.4
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.21.4
tomli==2.2.1
tomlkit==0.13.3
torch==2.8.0
torchinfo==1.8.0
torchvision==0.23.0
tqdm==4.67.1
transformers==4.55.0
triton==3.4.0
tsfeatures==0.4.5
tsfresh==0.21.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.37.0
Werkzeug==3.1.3
zipp==3.23.0
